requirements.txt
langchain
openai
chromadb
ragas[eval]
datasets
tiktoken
langgraph
beautifulsoup4
langchain-community
langchain-openai
gradio
python-dotenv
rank_bm25